wire up game map to GameViewer
